Moved rt loader to io module
