+++ brainbox/io/io.py
@@ -3,6 +3,9 @@ import numpy as np
 # (Previously required `os.path` to get file info before memmapping)
 # import os.path as op
 from ibllib.io import spikeglx
+from ibllib.io.extractors.training_wheel import extract_wheel_moves
+from ibllib.io.extractors.training_trials import FirstMovementTimes
+from oneibl.one import ONE
 
 
 def extract_waveforms(ephys_file, ts, ch, t=2.0, sr=30000, n_ch_probe=385, dtype='int16',
@@ -132,3 +135,45 @@ def extract_waveforms(ephys_file, ts, ch, t=2.0, sr=30000, n_ch_probe=385, dtype
         waveforms -= noise_s[None, None, :]
 
     return waveforms
+
+
+def load_wheel_reaction_times(eid, one=None):
+    """
+    Return the calculated reaction times for session.  Reaction times are defined as the time
+    between the go cue (onset tone) and the onset of the first substantial wheel movement.   A
+    movement is considered sufficiently large if its peak amplitude is at least 1/3rd of the
+    distance to threshold (~0.1 radians).
+
+    Negative times mean the onset of the movement occurred before the go cue.  Nans may occur if
+    there was no detected movement withing the period, or when the goCue_times or feedback_times
+    are nan.
+
+    Parameters
+    ----------
+    eid : str
+        Session UUID
+    one : oneibl.ONE
+        An instance of ONE for loading data.  If None a new one is instantiated using the defaults.
+
+    Returns
+    ----------
+    array-like
+        reaction times
+    """
+    if one is None:
+        one = ONE()
+
+    trials = one.load_object(eid, 'trials')
+    # If already extracted, load and return
+    if trials and 'firstMovement_times' in trials:
+        return trials['firstMovement_times'] - trials['goCue_times']
+    # Otherwise load wheelMoves object and calculate
+    moves = one.load_object(eid, 'wheelMoves')
+    # Re-extract wheel moves if necessary
+    if not moves or 'peakAmplitude' not in moves:
+        wheel = one.load_object(eid, 'wheel')
+        moves = extract_wheel_moves(wheel['timestamps'], wheel['position'])
+    assert trials and moves, 'unable to load trials and wheelMoves data'
+    firstMove_times, is_final_movement, ids = \
+        FirstMovementTimes.extract_first_movement_times(moves, trials)
+    return firstMove_times - trials['goCue_times']
